Server/requirements.txt
fastapi>=0.114,<0.116
uvicorn>=0.30,<0.32
pydantic>=2.7,<3

# optional but useful
numpy>=1.26,<3
pandas>=2.2,<3

# testing
pytest>=8.0,<9
httpx>=0.27,<0.28

joblib>=1.3,<2
scikit-learn>=1.4,<2   # only if you'll load sklearn pickles
python-dotenv>=1.0,<2  # optional, for .env config

tensorflow-cpu>=2.16,<3
numpy>=1.26
pillow>=10.0
python-multipart
